move match_domains to util
